staffstuddash/backend/requirements.txt
fastapi
pydantic
groq
python-dotenv